+++ src/main.jsx
@@ -32,24 +32,6 @@ function logout() {
     userStorage.reset();
 }
 
-function login() {
-    const username = document.getElementById("username");
-
-    let user = {
-        username: username.value,
-        email: "",
-        bio: "",
-    };
-
-    try {
-        localStorage.setItem("user", JSON.stringify(user));
-        globalStore.setState({ currentUser: user, loggedIn: true });
-        router.push("/profile");
-    } catch (error) {
-        globalStore.setState({ error });
-    }
-}
-
 function handleError(error) {
     globalStore.setState({ error });
 }
@@ -57,7 +39,6 @@ function handleError(error) {
 // 초기화 함수
 function render() {
     const $root = document.querySelector("#root");
-    const eventDelegator = initializeEventDelegator(); // 이벤트 위임 관리 객체 생성
     const targetPage = router.getTarget();
 
     try {
@@ -104,11 +85,6 @@ function main() {
     addEvent("click", "#error-boundary", (e) => {
         e.preventDefault();
         globalStore.setState({ error: null });
-    });
-
-    addEvent("submit", "#login-form", (e) => {
-        e.preventDefault();
-        login();
     });
 
     render();
